🦄 refactor(common): use self db
