--- common/repo/test/res/ip.yml
@@ -1,65 +1,0 @@
-#
-# Licensed under the GNU General Public License v. 3 (the "License");
-# You may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.gnu.org/licenses/gpl-3.0.html
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Copyright (C) 2023-2023 xqyjlj<xqyjlj@126.com>
-#
-# @author      xqyjlj
-# @file        ip.yml
-#
-# Change Logs:
-# Date           Author       Notes
-# ------------   ----------   -----------------------------------------------
-# 2023-06-18     xqyjlj       initial version
-#
-
-Input-Std:
-  chal_gpio_pull_t:
-    - CHAL_GPIO_PULL_NO
-    - CHAL_GPIO_PULL_UP
-    - CHAL_GPIO_PULL_DOWN
-  chal_gpio_mode_t:
-    - CHAL_GPIO_MODE_INPUT
-Output-Std:
-  chal_gpio_level_t:
-    - CHAL_GPIO_LEVEL_LOW
-    - CHAL_GPIO_LEVEL_HIGH
-  chal_gpio_speed_t:
-    - CHAL_GPIO_SPEED_10MHz
-    - CHAL_GPIO_SPEED_20MHz
-    - CHAL_GPIO_SPEED_50MHz
-  chal_gpio_mode_t:
-    - CHAL_GPIO_MODE_OUTPUT
-  chal_gpio_output_type_t:
-    - CHAL_GPIO_OUTPUT_PP
-    - CHAL_GPIO_OUTPUT_OD
-Analog-Std:
-  chal_gpio_mode_t:
-    - CHAL_GPIO_MODE_ANALOG
-Eventout-Std:
-  chal_gpio_pull_t:
-    - CHAL_GPIO_PULL_NO
-    - CHAL_GPIO_PULL_UP
-    - CHAL_GPIO_PULL_DOWN
-  chal_gpio_speed_t:
-    - CHAL_GPIO_SPEED_10MHz
-    - CHAL_GPIO_SPEED_20MHz
-    - CHAL_GPIO_SPEED_50MHz
-  chal_gpio_mode_t:
-    - CHAL_GPIO_MODE_ALTERNATE
-  chal_gpio_output_type_t:
-    - CHAL_GPIO_OUTPUT_PP
-Exti-Std:
-  chal_gpio_pull_t:
-    - CHAL_GPIO_PULL_NO
-    - CHAL_GPIO_PULL_UP
-    - CHAL_GPIO_PULL_DOWN
