--- common/repo/test/res/map.yml
@@ -1,163 +1,0 @@
-# Licensed under the Apache License, Version 2.0 (the "License");
-# You may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Copyright (C) 2022-2023 xqyjlj<xqyjlj@126.com>
-#
-# @author      xqyjlj
-# @file        gpio.yml
-#
-# Change Logs:
-# Date           Author       Notes
-# ------------   ----------   -----------------------------------------------
-# 2023-01-07     xqyjlj       initial version
-#
-
-groups:
-  chal_gpio_level_t:
-    comment:
-      zh_CN: GPIO电平类型
-      en:
-    values:
-      CHAL_GPIO_LEVEL_LOW:
-        comment:
-          zh_CN: 低电平
-          en:
-      CHAL_GPIO_LEVEL_HIGH:
-        comment:
-          zh_CN: 高电平
-          en:
-  GPIO_MODE_T:
-    comment:
-      zh_CN: GPIO模式类型
-      en:
-    values:
-      GPIO_MODE_ANALOG:
-        comment:
-          zh_CN: 模拟模式
-          en:
-      GPIO_MODE_IN_FLOATING:
-        comment:
-          zh_CN: 浮空输入
-          en:
-      GPIO_MODE_IN_PD:
-        comment:
-          zh_CN: 下拉输入
-          en:
-      GPIO_MODE_IN_PU:
-        comment:
-          zh_CN: 上拉输入
-          en:
-      GPIO_MODE_OUT_PP:
-        comment:
-          zh_CN: 推挽输出
-          en:
-      GPIO_MODE_OUT_OD:
-        comment:
-          zh_CN: 开漏输出
-          en:
-      GPIO_MODE_AF_PP:
-        comment:
-          zh_CN: 推挽复用
-          en:
-      GPIO_MODE_AF_OD:
-        comment:
-          zh_CN: 开漏复用
-          en:
-  chal_gpio_output_type_t:
-    comment:
-      zh_CN: GPIO输出类型
-      en:
-    values:
-      CHAL_GPIO_OUTPUT_PP:
-        comment:
-          zh_CN: 推挽输出
-          en:
-      CHAL_GPIO_OUTPUT_OD:
-        comment:
-          zh_CN: 开漏输出
-          en:
-  chal_gpio_pull_t:
-    comment:
-      zh_CN: GPIO钳位类型
-      en:
-    values:
-      CHAL_GPIO_PULL_DOWN:
-        comment:
-          zh_CN: 下拉
-          en:
-      CHAL_GPIO_PULL_UP:
-        comment:
-          zh_CN: 上拉
-          en:
-  chal_gpio_speed_t:
-    comment:
-      zh_CN: GPIO速度类型
-      en:
-    values:
-      CHAL_GPIO_SPEED_10MHz:
-        comment:
-          zh_CN: 10MHz
-          en:
-      CHAL_GPIO_SPEED_20MHz:
-        comment:
-          zh_CN: 20MHz
-          en:
-      CHAL_GPIO_SPEED_50MHz:
-        comment:
-          zh_CN: 50MHz
-          en:
-properties:
-  chal_gpio_level_t:
-    display_name:
-      zh_CN: 电平
-      en:
-    description:
-      zh_CN: GPIO-电平
-      en:
-    category: GPIO
-    readonly: false
-  chal_gpio_mode_t:
-    display_name:
-      zh_CN: 模式
-      en:
-    description:
-      zh_CN: GPIO-模式
-      en:
-    category: GPIO
-    readonly: false
-  chal_gpio_output_type_t:
-    display_name:
-      zh_CN: 输出类型
-      en:
-    description:
-      zh_CN: GPIO-输出类型
-      en:
-    category: GPIO
-    readonly: false
-  chal_gpio_pull_t:
-    display_name:
-      zh_CN: 钳位
-      en:
-    description:
-      zh_CN: GPIO-钳位
-      en:
-    category: GPIO
-    readonly: false
-  chal_gpio_speed_t:
-    display_name:
-      zh_CN: 速度
-      en:
-    description:
-      zh_CN: GPIO-速度
-      en:
-    category: GPIO
-    readonly: false
