--- common/repo/test/res/repository.yml
@@ -1,86 +1,0 @@
-# Licensed under the Apache License, Version 2.0 (the "License");
-# You may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Copyright (C) 2022-2023 xqyjlj<xqyjlj@126.com>
-#
-# @author      xqyjlj
-# @file        repository.yml
-#
-# Change Logs:
-# Date           Author       Notes
-# ------------   ----------   -----------------------------------------------
-# 2023-03-09     csplinkbot   auto update
-#
-chips:
-  Geehy:
-    APM32F1:
-      APM32F103:
-        APM32F103C8T6:
-          core: Arm Cortex-M3
-          current:
-            lowest: 1.9
-            run: 421.0
-          flash: 64
-          frequency: 96
-          io: 37
-          package: LQFP48
-          peripherals:
-            ADC 12-bit: 21
-            CAN: 1
-            DAC 12-bit: 2
-            FSMC: 1
-            I2C: 2
-            I2S: 2
-            RTC: 1
-            SDIO: 1
-            SPI: 3
-            Timer 16-bit: 8
-            UART: 2
-            USART: 3
-            USB Device: 1
-          ram: 20
-          temperature:
-            max: 105.0
-            min: -40.0
-          voltage:
-            max: 3.6
-            min: 2.0
-        APM32F103ZET6:
-          core: Arm Cortex-M3
-          current:
-            lowest: 1.9
-            run: 421.0
-          flash: 512
-          frequency: 96
-          io: 112
-          package: LQFP144
-          peripherals:
-            ADC 12-bit: 21
-            CAN: 1
-            DAC 12-bit: 2
-            FSMC: 1
-            I2C: 2
-            I2S: 2
-            RTC: 1
-            SDIO: 1
-            SPI: 3
-            Timer 16-bit: 8
-            UART: 2
-            USART: 3
-            USB Device: 1
-          ram: 128
-          temperature:
-            max: 105.0
-            min: -40.0
-          voltage:
-            max: 3.6
-            min: 2.0
