+++ common/repo/test/res/res.qrc
@@ -1,9 +1,4 @@
 <RCC>
     <qresource prefix="/">
-        <file>geehy/apm32f103zet6.yml</file>
-        <file>ip.yml</file>
-        <file>map.yml</file>
-        <file>pinout.yml</file>
-        <file>repository.yml</file>
     </qresource>
 </RCC>
